+++ src/api/tests/Forward.java
@@ -8,7 +8,7 @@ import api.classes.TurtleCommand;
 public class Forward extends TurtleCommand {
 
     @Override
-    public void generateUpdate (float distance, Queue<StateUpdate> stateQueue) {
+    public void generateUpdate (double distance, Queue<StateUpdate> stateQueue) {
 
     }
