+++ src/commandParsing/mathCommandParsing/NaturalLog.java
@@ -11,8 +11,8 @@ import drawableobject.DrawableObject;
 public class NaturalLog extends OneInputFloatCommandParser {
 	
 	@Override
-	protected float operateOnComponents(List<Float> components,	Queue<DrawableObject> objectQueue) throws RunTimeDivideByZeroException {
-		return (float) Math.log(components.get(0));
+	protected double operateOnComponents(List<Float> components,	Queue<DrawableObject> objectQueue) throws RunTimeDivideByZeroException {
+		return (double) Math.log(components.get(0));
 	}
 
 }
