+++ src/tests/commandTests/doubleInputFloatTests/DifferenceTests.java
@@ -20,7 +20,7 @@ public class DifferenceTests extends CommandTester{
 		setUpCommands("- 50 50");
 		
 		CommandParser parser = createCommand();
-		float f = parser.parse(commands, objectQueue);
+		double f = parser.parse(commands, objectQueue);
 		assertTrue(f == 0);
 	}
 
@@ -30,7 +30,7 @@ public class DifferenceTests extends CommandTester{
 		setUpCommands("- 50.0 50.0");
 		CommandParser parser = createCommand();
 
-		float f = parser.parse(commands, objectQueue);
+		double f = parser.parse(commands, objectQueue);
 
 		assertTrue(f == 0);
 	}
@@ -41,7 +41,7 @@ public class DifferenceTests extends CommandTester{
 		setUpCommands("- 50 - 50 50");
 		CommandParser parser = createCommand();
 
-		float f = parser.parse(commands, objectQueue);
+		double f = parser.parse(commands, objectQueue);
 
 		assertTrue(f == 50);
 	}
@@ -52,7 +52,7 @@ public class DifferenceTests extends CommandTester{
 		setUpCommands("- 50 - 50 - 50 - 50 - 50 - 50 - 50 50");
 		CommandParser parser = createCommand();
 
-		float f = parser.parse(commands, objectQueue);
+		double f = parser.parse(commands, objectQueue);
 
 		assertTrue(f == 0);
 	}
@@ -64,7 +64,7 @@ public class DifferenceTests extends CommandTester{
 		CommandParser parser = createCommand();
 
 		try {
-			float f = parser.parse(commands, objectQueue);
+			double f = parser.parse(commands, objectQueue);
 		} catch (SLOGOException e) {
 			assertTrue(e.generateErrorMessage().getParameters().values().contains("Error parsing following string: " + "50.." + ". Incorrect syntax."));
 		}
